first_web_gui and metro4-theme where added
