+++ Web/Gui/first_web_gui/static/styles.css
@@ -1,0 +1,1 @@
+/* Custom CSS */
