+++ Web/Gui/metro4-theme/static/styles.css
@@ -1,0 +1,1 @@
+/* Custom CSS */
